requirements.txt
numpy
opencv-python-headless
torchvision
albumentations==1.3.1
Flask==2.0.3


